fix(UI): Display correct round number
Previously, the main header displayed the correct, current round number,
but the round number in "Make a selection to play round ____" was
incorrect. The round number displayed there was the same as the round
number displayed in the main header, whereas it should be one more than
the main header. This is because it represents the next round, not the
current round. Correct this and include a special case message when the
player is about to play the last round.